validate-consumer/requirements.txt
kafka-python==2.0.2
aiokafka==0.7.1
uvicorn==0.13.3
flake8==3.9.0
python-dotenv==0.19.0
sqlalchemy==1.4.51
mysql-connector-python==8.2.0